ar71xx: factory image generation for WNDR4300
Add factory image generation using UBI/UBIFS. The rootfile is
mounted rw, no overlay support yet.

Just in front of the rootfs (still on the kernel partition),
a fake header needs to be in place to make stock U-Boot happy.
This fake header tells the system rootfs is 4 bytes long. The
first for bytes of the rootfs is an UBI magic (ASCII UBI#).